Feat: Added flag --no-write-check on smb shares

+++ cme/protocols/smb.py
@@ -867,15 +867,16 @@ class smb(connection):
                 self.logger.debug(f"Error checking READ access on share: {error}")
                 pass
 
-            try:
-                self.conn.createDirectory(share_name, temp_dir)
-                self.conn.deleteDirectory(share_name, temp_dir)
-                write = True
-                share_info["access"].append("WRITE")
-            except SessionError as e:
-                error = get_error_string(e)
-                self.logger.debug(f"Error checking WRITE access on share: {error}")
-                pass
+            if not self.args.no_write_check:
+                try:
+                    self.conn.createDirectory(share_name, temp_dir)
+                    self.conn.deleteDirectory(share_name, temp_dir)
+                    write = True
+                    share_info["access"].append("WRITE")
+                except SessionError as e:
+                    error = get_error_string(e)
+                    self.logger.debug(f"Error checking WRITE access on share: {error}")
+                    pass
 
             permissions.append(share_info)
